+++ src/main/java/org/sesac/slopedbe/auth/model/dto/JwtResponse.java
@@ -1,4 +1,4 @@
-package org.sesac.slopedbe.auth.model.vo;
+package org.sesac.slopedbe.auth.model.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
